add /style for css files, moved getwitz to service
