+++ client/src/components/WitzePanel.component.tsx
@@ -1,10 +1,10 @@
-import GetWitz from "./GetWitz"
+import GetWitz from "../services/GetWitz.service";
 
 function WitzePanel() {
   return (
     <div>
       <h2>Der Witz des Tages</h2>
-        <GetWitz/>
+      <GetWitz />
     </div>
   );
 }
