+++ client/src/style/Login.css
@@ -1,0 +1,3 @@
+/*
+* wird auf alles in login.component.tsx angewendet außer was untergeordetes  
+*/
