+++ client/src/style/MainPage.css
@@ -35,3 +35,36 @@ wird explizit definiert
     -webkit-border-radius: 50%;
     border-radius: 50%;
   }
+
+  .navbar-brand {
+    background-color: #2c3e50;
+    border-color: #2c3e50;
+  }
+
+  .nav-item {
+    background-color: #2c3e50;
+    border-color: #2c3e50;
+  }
+
+  .navbar-nav ml-auto {
+    background-color: #2c3e50;
+    border-color: #2c3e50;
+  }
+
+  .container mt-3 {
+    background-color: #2c3e50;
+    border-color: #2c3e50;
+  }
+
+  .nav-link {
+    color: #fff;
+  }
+
+  .navbar-nav mr-auto {
+    background-color: #2c3e50;
+    border-color: #2c3e50;
+  }
+
+  .navbar-brand {
+    color: #fff;
+  }
