+++ client/src/style/Register.css
@@ -1,0 +1,3 @@
+/*
+* wird auf alles in register.component.tsx angewendet außer was untergeordetes  
+*/
