+++ client/src/style/Userboard.css
@@ -1,0 +1,3 @@
+/*
+* wird auf alles in userboard.component.tsx angewendet außer was untergeordetes  
+*/
