+++ client/src/style/WitzePanel.css
@@ -1,0 +1,3 @@
+/*
+* wird auf alles in WitzePanel.component.tsx angewendet außer was untergeordetes  
+*/
